requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.6
gunicorn==22.0.0
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
PyJWT==2.8.0
pymongo==4.7.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.2
python-ulid==2.5.0
requests==2.32.2
simple-websocket==1.0.0
urllib3==2.2.1
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
